+++ test/java/com/rapleaf/jack/BaseDatabaseModelTestCase.java
@@ -106,7 +106,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
   
   public void testFindEmptySet() throws Exception {
     IUserPersistence users = dbs.getDatabase1().users();
-    Set<User> foundValues = users.find(new HashSet<Integer>());
+    Set<User> foundValues = users.find(new HashSet<Long>());
     assertEquals(0, foundValues.size());
   }
   
@@ -122,7 +122,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
 
     users.clearCacheById(bryand.getId());
     users.clearCacheById(notBryand.getId());
-    Set<Integer> keysToSearch = new HashSet<Integer>();
+    Set<Long> keysToSearch = new HashSet<Long>();
     keysToSearch.add(bryand.getId());
     keysToSearch.add(notBryand.getId());
     Set<User> foundValues = users.find(keysToSearch);
@@ -182,7 +182,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
     User notBryand = users.create("notBryand", t0, 3, t1, t2, "another relatively long string", someBinary, 1.2d, true);
     users.create("unwanted", t0, 0, t1, t2, "yet another relatively long string", someBinary, 1.2d, true);
 
-    Set<Integer> keysToSearch = new HashSet<Integer>();
+    Set<Long> keysToSearch = new HashSet<Long>();
     keysToSearch.add(bryand.getId());
     keysToSearch.add(notBryand.getId());
     Set<User> foundValues = users.find(keysToSearch);
@@ -281,15 +281,15 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
 
   public void testFindAllByForeignKeyFromSet() throws Exception {
     ICommentPersistence comments = dbs.getDatabase1().comments();
-    int userId = 1;
-    int otherUserId = 2;
-    Comment c1 = comments.create("comment1", userId, 1, 1);
-    Comment c2 = comments.create("comment2", userId, 1, 1);
-    Comment c3 = comments.create("comment3", userId, 1, 1);
-    Comment c4 = comments.create("comment4", otherUserId, 1, 1);
+    Long userId = 1L;
+    Long otherUserId = 2L;
+    Comment c1 = comments.create("comment1", userId.intValue(), 1, 1);
+    Comment c2 = comments.create("comment2", userId.intValue(), 1, 1);
+    Comment c3 = comments.create("comment3", userId.intValue(), 1, 1);
+    Comment c4 = comments.create("comment4", otherUserId.intValue(), 1, 1);
     Comment c5 = comments.create("comment5", 3, 1, 1);
 
-    Set<Integer> commenterIds = new HashSet<Integer>();
+    Set<Long> commenterIds = new HashSet<Long>();
     commenterIds.add(userId);
     commenterIds.add(otherUserId);
     Set<Comment> userComments = comments.findAllByForeignKey("commenter_id", commenterIds);
@@ -386,7 +386,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
     User u1 = users.create("bryand", System.currentTimeMillis(), 5, System.currentTimeMillis() + 10, System.currentTimeMillis() + 20, "this is a relatively long string", new byte[]{5, 4, 3, 2, 1}, 1.2d, true);
 
     IPostPersistence posts = dbs.getDatabase1().posts();
-    Post p1 = posts.create("title", System.currentTimeMillis(), u1.getId());
+    Post p1 = posts.create("title", System.currentTimeMillis(), (int) u1.getId());
     assertEquals(u1, p1.getUser());
   }
 
@@ -395,7 +395,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
     User u1 = users.create("bryand", System.currentTimeMillis(), 5, System.currentTimeMillis() + 10, System.currentTimeMillis() + 20, "this is a relatively long string", new byte[]{5, 4, 3, 2, 1}, 1.2d, true);
 
     IImagePersistence images = dbs.getDatabase1().images();
-    Image image = images.create(u1.getId());
+    Image image = images.create((int) u1.getId());
     assertEquals(u1, image.getUser());
   }
 
@@ -404,9 +404,9 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
     User u1 = users.create("bryand", System.currentTimeMillis(), 5, System.currentTimeMillis() + 10, System.currentTimeMillis() + 20, "this is a relatively long string", new byte[]{5, 4, 3, 2, 1}, 1.2d, true);
 
     IPostPersistence posts = dbs.getDatabase1().posts();
-    Post p1 = posts.create("title1", System.currentTimeMillis(), u1.getId());
-    Post p2 = posts.create("title2", System.currentTimeMillis(), u1.getId());
-    Post p3 = posts.create("title3", System.currentTimeMillis(), u1.getId());
+    Post p1 = posts.create("title1", System.currentTimeMillis(), (int) u1.getId());
+    Post p2 = posts.create("title2", System.currentTimeMillis(), (int) u1.getId());
+    Post p3 = posts.create("title3", System.currentTimeMillis(), (int) u1.getId());
 
     assertEquals(new HashSet<Post>(Arrays.asList(p1, p2, p3)), u1.getPosts());
   }
@@ -433,7 +433,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
   public void testDelete() throws Exception {
     IPostPersistence posts = dbs.getDatabase1().posts();
     Post post = posts.create(null, 10L, 1);
-    int id = post.getId();
+    long id = post.getId();
     posts.delete(id);
     assertNull(posts.find(id));
   }
@@ -441,7 +441,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
   public void testSave() throws Exception {
     IPostPersistence posts = dbs.getDatabase1().posts();
     Post post = posts.create(null, 10L, 1);
-    int id = post.getId();
+    long id = post.getId();
     post.setPostedAtMillis(20L);
     dbs.getDatabase1().posts().save(post);
     assertEquals(Long.valueOf(20), posts.find(id).getPostedAtMillis());
@@ -478,7 +478,7 @@ public abstract class BaseDatabaseModelTestCase extends TestCase {
     assertTrue(found.contains(u1));
     assertTrue(found.contains(u2));
 
-    found = users.find(new HashSet<Integer>(Arrays.asList(u1.getId())), new HashMap<Enum, Object>(){{put(User._Fields.num_posts, 2);}});
+    found = users.find(new HashSet<Long>(Arrays.asList(u1.getId())), new HashMap<Enum, Object>(){{put(User._Fields.num_posts, 2);}});
     assertEquals(1, found.size());
     assertTrue(found.contains(u1));
   }
